test_analyzer/requirements.txt
azure-kusto-data==3.1.3
azure-kusto-ingest==3.1.3
defusedxml==0.7.1
setuptools-rust==1.1.2
pandas==2.0.3
crontab==1.0.1